exercise-round-7/template_storage/requirements.txt
google-cloud-storage
google-cloud-vision
google-api-core
grpcio